Move goblin verifier
