+++ barretenberg/cpp/src/barretenberg/aztec_ivc/CMakeLists.txt
@@ -1,1 +1,1 @@
-barretenberg_module(aztec_ivc goblin stdlib_protogalaxy_verifier)
+barretenberg_module(aztec_ivc stdlib_protogalaxy_verifier stdlib_goblin_verifier)
