+++ barretenberg/cpp/src/barretenberg/benchmark/goblin_bench/CMakeLists.txt
@@ -1,1 +1,1 @@
-barretenberg_module(goblin_bench eccvm translator_vm stdlib_honk_recursion stdlib_sha256 crypto_merkle_tree)
+barretenberg_module(goblin_bench eccvm)
