+++ barretenberg/cpp/src/barretenberg/client_ivc/CMakeLists.txt
@@ -1,1 +1,1 @@
-barretenberg_module(client_ivc goblin stdlib_protogalaxy_verifier)
+barretenberg_module(client_ivc stdlib_protogalaxy_verifier stdlib_goblin_verifier)
