+++ barretenberg/cpp/src/barretenberg/stdlib/CMakeLists.txt
@@ -1,6 +1,7 @@
 add_subdirectory(hash)
 add_subdirectory(commitment)
 add_subdirectory(encryption)
+add_subdirectory(goblin_verifier)
 add_subdirectory(primitives)
 add_subdirectory(honk_recursion)
 add_subdirectory(transcript)
