+++ barretenberg/cpp/src/barretenberg/stdlib/goblin_verifier/CMakeLists.txt
@@ -1,0 +1,5 @@
+barretenberg_module(
+    stdlib_goblin_verifier
+    stdlib_circuit_builders
+    goblin
+)
