+++ barretenberg/cpp/src/barretenberg/stdlib/goblin_verifier/goblin_recursive_verifier.cpp
@@ -1,4 +1,4 @@
-#include "barretenberg/stdlib/honk_recursion/verifier/goblin_recursive_verifier.hpp"
+#include "barretenberg/stdlib/goblin_verifier/goblin_recursive_verifier.hpp"
 
 namespace bb::stdlib::recursion::honk {
