+++ barretenberg/cpp/src/barretenberg/stdlib/goblin_verifier/goblin_recursive_verifier.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "barretenberg/eccvm_recursion/eccvm_recursive_verifier.hpp"
 #include "barretenberg/goblin/goblin.hpp"
-#include "barretenberg/stdlib/honk_recursion/verifier/merge_recursive_verifier.hpp"
+#include "barretenberg/stdlib/goblin_verifier/merge_recursive_verifier.hpp"
 #include "barretenberg/translator_vm_recursion/translator_recursive_verifier.hpp"
 
 namespace bb::stdlib::recursion::honk {
