+++ barretenberg/cpp/src/barretenberg/stdlib/honk_recursion/verifier/client_ivc_recursive_verifier.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "barretenberg/client_ivc/client_ivc.hpp"
+#include "barretenberg/stdlib/goblin_verifier/goblin_recursive_verifier.hpp"
 #include "barretenberg/stdlib/honk_recursion/verifier/decider_recursive_verifier.hpp"
-#include "barretenberg/stdlib/honk_recursion/verifier/goblin_recursive_verifier.hpp"
 
 namespace bb::stdlib::recursion::honk {
 class ClientIVCRecursiveVerifier {
